Added SlimuxSendKeys function
with some wrapper commands for easy mapping. This is similar to
SlimuxSendCommand, but it is a bit more low level as it doesn't insert carriage
return(Enter) automatically and enables sending special keys to the pane(such as
ctrl+c for SIGINT).

Also modified SlimuxShellLast so it will prompt for a command if none was sent
before.